+++ src/main/java/org/opensearch/knn/common/FieldInfoExtractor.java
@@ -1,0 +1,50 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package org.opensearch.knn.common;
+
+import lombok.experimental.UtilityClass;
+import org.apache.lucene.index.FieldInfo;
+import org.opensearch.knn.index.VectorDataType;
+import org.opensearch.knn.index.engine.KNNEngine;
+import org.opensearch.knn.indices.ModelMetadata;
+
+import static org.opensearch.knn.common.KNNConstants.MODEL_ID;
+import static org.opensearch.knn.indices.ModelUtil.getModelMetadata;
+
+@UtilityClass
+public class FieldInfoExtractor {
+
+    /**
+     * Extracts KNNEngine from FieldInfo
+     * @param field {@link FieldInfo}
+     * @return {@link KNNEngine}
+     */
+    public static KNNEngine extractKNNEngine(final FieldInfo field) {
+        final ModelMetadata modelMetadata = getModelMetadata(field.attributes().get(MODEL_ID));
+        if (modelMetadata != null) {
+            return modelMetadata.getKnnEngine();
+        }
+        final String engineName = field.attributes().getOrDefault(KNNConstants.KNN_ENGINE, KNNEngine.DEFAULT.getName());
+        return KNNEngine.getEngine(engineName);
+    }
+
+    /**
+     * Extracts VectorDataType from FieldInfo
+     * @param field {@link FieldInfo}
+     * @return {@link VectorDataType}
+     */
+    public static VectorDataType extractVectorDataType(final FieldInfo field) {
+        String vectorDataTypeString = field.getAttribute(KNNConstants.VECTOR_DATA_TYPE_FIELD);
+        if (vectorDataTypeString == null) {
+            final ModelMetadata modelMetadata = getModelMetadata(field.attributes().get(MODEL_ID));
+            if (modelMetadata != null) {
+                VectorDataType vectorDataType = modelMetadata.getVectorDataType();
+                vectorDataTypeString = vectorDataType == null ? null : vectorDataType.getValue();
+            }
+        }
+        return vectorDataTypeString != null ? VectorDataType.get(vectorDataTypeString) : VectorDataType.DEFAULT;
+    }
+}