+++ src/test/java/org/opensearch/knn/index/memory/NativeMemoryAllocationTests.java
@@ -14,6 +14,7 @@ package org.opensearch.knn.index.memory;
 import com.google.common.collect.ImmutableMap;
 import lombok.SneakyThrows;
 import org.opensearch.knn.KNNTestCase;
+import org.opensearch.knn.TestUtils;
 import org.opensearch.knn.common.KNNConstants;
 import org.opensearch.knn.index.util.IndexUtil;
 import org.opensearch.knn.index.VectorDataType;
@@ -56,7 +57,7 @@ public class NativeMemoryAllocationTests extends KNNTestCase {
         }
         Map<String, Object> parameters = ImmutableMap.of(KNNConstants.SPACE_TYPE, SpaceType.DEFAULT.getValue());
         long vectorMemoryAddress = JNICommons.storeVectorData(0, vectors, numVectors * dimension);
-        JNIService.createIndex(ids, vectorMemoryAddress, dimension, path, parameters, knnEngine);
+        TestUtils.createIndex(ids, vectorMemoryAddress, dimension, path, parameters, knnEngine);
 
         // Load index into memory
         long memoryAddress = JNIService.loadIndex(path, parameters, knnEngine);
@@ -117,7 +118,7 @@ public class NativeMemoryAllocationTests extends KNNTestCase {
             VectorDataType.BINARY.getValue()
         );
         long vectorMemoryAddress = JNICommons.storeByteVectorData(0, vectors, numVectors * dataLength);
-        JNIService.createIndex(ids, vectorMemoryAddress, dimension, path, parameters, knnEngine);
+        TestUtils.createIndex(ids, vectorMemoryAddress, dimension, path, parameters, knnEngine);
 
         // Load index into memory
         long memoryAddress = JNIService.loadIndex(path, parameters, knnEngine);